Pull editable logic into separate service
